feature: Rename AirframeException to DesignException and update references
