+++ ai-core/src/main/scala/wvlet/ai/design/DesignException.scala
@@ -18,7 +18,7 @@ import wvlet.ai.util.SourceCode
 
 import scala.language.existentials
 
-trait AirframeException extends Exception:
+trait DesignException extends Exception:
   self =>
 
   /**
@@ -27,28 +27,27 @@ trait AirframeException extends Exception:
   def getCode: String           = this.getClass.getSimpleName
   override def toString: String = getMessage
 
-object AirframeException:
-  case class MISSING_SESSION(cl: Class[?]) extends AirframeException:
+object DesignException:
+  case class MISSING_SESSION(cl: Class[?]) extends DesignException:
     override def getMessage: String =
       s"[$getCode] Session is not found inside ${cl}. You may need to define ${cl} as a trait or implement DISupport to inject the current Session."
 
-  case class CYCLIC_DEPENDENCY(deps: List[Surface], sourceCode: SourceCode)
-      extends AirframeException:
+  case class CYCLIC_DEPENDENCY(deps: List[Surface], sourceCode: SourceCode) extends DesignException:
     override def getMessage: String =
       s"[$getCode] ${deps.reverse.mkString(" -> ")} at ${sourceCode}"
 
   case class MISSING_DEPENDENCY(stack: List[Surface], sourceCode: SourceCode)
-      extends AirframeException:
+      extends DesignException:
     override def getMessage: String =
       s"[$getCode] Binding for ${stack.head} at ${sourceCode} is not found: ${stack.mkString(
           " <- "
         )}"
 
-  case class SHUTDOWN_FAILURE(cause: Throwable) extends AirframeException:
+  case class SHUTDOWN_FAILURE(cause: Throwable) extends DesignException:
     override def getMessage: String =
       s"[${getCode}] Failure at session shutdown: ${cause.getMessage}"
 
-  case class MULTIPLE_SHUTDOWN_FAILURES(causes: List[Throwable]) extends AirframeException:
+  case class MULTIPLE_SHUTDOWN_FAILURES(causes: List[Throwable]) extends DesignException:
     override def getMessage: String =
       s"[${getCode}] Multiple failures occurred during session shutdown:\n${causes
           .map(x => s"  - ${x.getMessage}")
