+++ ai-core/src/test/scala/wvlet/ai/design/DependencyTest.scala
@@ -1,6 +1,6 @@
 package wvlet.ai.design
 
-import wvlet.ai.design.AirframeException.MISSING_DEPENDENCY
+import wvlet.ai.design.DesignException.MISSING_DEPENDENCY
 import wvlet.ai.design.Design
 import wvlet.airspec.AirSpec
